refactor: Rename node-resize components to f-node-resize for consistency
