+++ projects/f-flow/src/f-draggable/f-node-resize/get-node-resize-restrictions/get-node-resize-restrictions.execution.ts
