+++ projects/f-flow/src/f-draggable/f-node-resize/get-node-resize-restrictions/i-node-resize-restrictions.ts
