+++ projects/f-flow/src/f-draggable/f-node-resize/providers.ts
@@ -1,5 +1,5 @@
-import { NODE_RESIZE_FINALIZE_PROVIDERS } from './node-resize-finalize';
-import { NodeResizePreparationExecution } from './node-resize-preparation';
+import { FNodeResizeFinalizeExecution } from './resize-finalize';
+import { FNodeResizePreparationExecution } from './resize-preparation';
 import { GetNodeResizeRestrictionsExecution } from './get-node-resize-restrictions';
 import { ApplyChildResizeRestrictionsExecution } from './apply-child-resize-restrictions';
 import { ApplyParentResizeRestrictionsExecution } from './apply-parent-resize-restrictions';
@@ -21,7 +21,7 @@ export const NODE_RESIZE_PROVIDERS = [
 
   GetNodeResizeRestrictionsExecution,
 
-  ...NODE_RESIZE_FINALIZE_PROVIDERS,
+  FNodeResizeFinalizeExecution,
 
-  NodeResizePreparationExecution,
+  FNodeResizePreparationExecution,
 ];
