+++ projects/f-flow/src/f-draggable/f-node-resize/resize-finalize/f-node-resize-finalize.execution.ts
@@ -1,16 +1,16 @@
 import { inject, Injectable } from '@angular/core';
-import { NodeResizeFinalizeRequest } from './node-resize-finalize.request';
+import { FNodeResizeFinalizeRequest } from './f-node-resize-finalize.request';
 import { FExecutionRegister, IExecution } from '@foblex/mediator';
 import { FDraggableDataContext } from '../../f-draggable-data-context';
-import { NodeResizeDragHandler } from '../node-resize.drag-handler';
+import { FNodeResizeDragHandler } from '../f-node-resize.drag-handler';
 
 @Injectable()
-@FExecutionRegister(NodeResizeFinalizeRequest)
-export class NodeResizeFinalizeExecution implements IExecution<NodeResizeFinalizeRequest, void> {
+@FExecutionRegister(FNodeResizeFinalizeRequest)
+export class FNodeResizeFinalizeExecution implements IExecution<FNodeResizeFinalizeRequest, void> {
 
   private _fDraggableDataContext = inject(FDraggableDataContext);
 
-  public handle(request: NodeResizeFinalizeRequest): void {
+  public handle(request: FNodeResizeFinalizeRequest): void {
     if (!this._isValid()) {
       return;
     }
@@ -21,7 +21,7 @@ export class NodeResizeFinalizeExecution implements IExecution<NodeResizeFinaliz
 
   private _isValid(): boolean {
     return this._fDraggableDataContext.draggableItems.some((x) =>
-      x instanceof NodeResizeDragHandler
+      x instanceof FNodeResizeDragHandler
     );
   }
 }
