+++ projects/f-flow/src/f-draggable/f-node-resize/resize-finalize/index.ts
@@ -1,0 +1,3 @@
+export * from './f-node-resize-finalize.execution';
+
+export * from './f-node-resize-finalize.request';
