--- projects/f-flow/src/f-draggable/node-resize/node-resize-finalize/index.ts
@@ -1,5 +1,0 @@
-export * from './providers';
-
-export * from './node-resize-finalize.execution';
-
-export * from './node-resize-finalize.request';
