--- projects/f-flow/src/f-draggable/node-resize/node-resize-finalize/providers.ts
@@ -1,6 +1,0 @@
-import { NodeResizeFinalizeExecution } from './node-resize-finalize.execution';
-
-export const NODE_RESIZE_FINALIZE_PROVIDERS = [
-
-  NodeResizeFinalizeExecution,
-];
