--- projects/f-flow/src/f-draggable/node-resize/node-resize-preparation/index.ts
@@ -1,3 +1,0 @@
-export * from './node-resize-preparation.execution';
-
-export * from './node-resize-preparation.request';
